Move resolve submodules into their parents
