+++ betty/data/aggregate/record/__init__.py
@@ -13,7 +13,7 @@ from betty.assertion import OptionalField, assert_mapping
 from betty.data import DataDefinition, OptionalDefinition, Sample, Samples
 from betty.data.aggregate import AggregateDefinition
 from betty.data.indicator.selector import Element
-from betty.locale.localizable.resolve import resolve_localizable
+from betty.locale.localizable import resolve_localizable
 from betty.portable import Portable, PortableData, PortablePorter, Porter
 
 if TYPE_CHECKING:
