+++ betty/document.py
@@ -29,7 +29,7 @@ from typing_extensions import override
 from betty.json.linked_data import LinkedDataDumpable
 from betty.locale.localize import DEFAULT_LOCALIZER, Localizer
 from betty.media_type.media_types import HTML
-from betty.plugin.resolve import ResolvableId, resolve_id
+from betty.plugin import ResolvableId, resolve_id
 from betty.portable import PortableMapping
 
 if TYPE_CHECKING:
