+++ betty/test_utils/jinja2.py
@@ -15,7 +15,7 @@ from lxml.html import document_fromstring
 from betty.app import App
 from betty.jinja2 import Environment
 from betty.json.schema import AllOf, Ref
-from betty.plugin.resolve import ResolvableId
+from betty.plugin import ResolvableId
 from betty.project import Project
 from betty.project.schema import ProjectSchema
