+++ betty/tests/extension/raspberry_mint/test_maps_ui.py
@@ -4,7 +4,7 @@ from typing_extensions import override
 
 from betty.extension import ExtensionDefinition
 from betty.extension.raspberry_mint import RaspberryMint
-from betty.plugin.resolve import ResolvableId
+from betty.plugin import ResolvableId
 from betty.test_utils.project.extension.maps import MapsTestBase
